+++ portal/src/index.js
@@ -5,16 +5,19 @@ const SEARCH_BAR_CSS_CLASS = 'search-bar';
 const BACKGROUND_CSS_CLASS = 'background';
 const TOP_PAGES_CSS_CLASS = 'top-pages';
 const SIDEBAR_CSS_CLASS = 'sidebar';
+const WIDGETS_CSS_CLASS = 'widgets';
 
 export function toggle_sidebar() {
     if (is_visible(SIDEBAR_CSS_CLASS)) {
         hide(MOBILE_HEADER_CSS_CLASS, TOP_PAGES_CSS_CLASS);
         hide(BACKGROUND_CSS_CLASS);
         hide(SIDEBAR_CSS_CLASS);
+        hide(WIDGETS_CSS_CLASS);
     } else {
         show(MOBILE_HEADER_CSS_CLASS, TOP_PAGES_CSS_CLASS, SEARCH_BAR_CSS_CLASS);
         show(BACKGROUND_CSS_CLASS);
         show(SIDEBAR_CSS_CLASS);
+        show(WIDGETS_CSS_CLASS);
     }
 }
 
@@ -26,6 +29,7 @@ export function toggle_search_bar() {
         show(MOBILE_HEADER_CSS_CLASS, SEARCH_BAR_CSS_CLASS, TOP_PAGES_CSS_CLASS);
         show(BACKGROUND_CSS_CLASS);
         hide(SIDEBAR_CSS_CLASS);
+        hide(WIDGETS_CSS_CLASS);
     }
 }
 
@@ -34,7 +38,8 @@ function is_visible(element_css_class) {
 }
 
 function show(element_css_class, to_show, to_hide) {
-    if (element_css_class === MOBILE_HEADER_CSS_CLASS) {
+    if (element_css_class === WIDGETS_CSS_CLASS) document.getElementsByClassName(element_css_class).item(0).classList.add('d-flex');
+    else if (element_css_class === MOBILE_HEADER_CSS_CLASS) {
         document.getElementsByClassName(to_hide).item(0).classList.remove('d-block');
         document.getElementsByClassName(to_show).item(0).classList.add('d-block');
     }
@@ -44,6 +49,7 @@ function show(element_css_class, to_show, to_hide) {
 
 function hide(element_css_class, to_hide) {
     if (element_css_class === MOBILE_HEADER_CSS_CLASS) document.getElementsByClassName(to_hide).item(0).classList.remove('d-block');
+    else if (element_css_class === WIDGETS_CSS_CLASS) document.getElementsByClassName(element_css_class).item(0).classList.remove('d-flex');
 
     document.getElementsByClassName(element_css_class).item(0).classList.add('d-none');
 }